Move dataset into processing
